% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_duplicates.R
\name{assert_duplicates}
\alias{assert_duplicates}
\title{Title}
\usage{
assert_duplicates(indata, ...)
}
\arguments{
\item{indata}{input dataset}

\item{...}{keys that you want to test uniqueness}
}
\value{
a
}
\description{
Title
}
\examples{
#' get_duplicates(iris,Species)
}
