% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{get_dup_test_data}
\alias{get_dup_test_data}
\title{Random Created data mimic P3 project}
\format{
A data frame with 27 rows and 8 variables:
\describe{
\item{pol_nbr}{policy number normally 7 digits}
\item{pol_eff_date}{policy effective date}
\item{premises_nbr}{Premises number}
\item{bldg_nbr}{building number}
\item{v1}{fake index1}
\item{v2}{fake index2}
\item{v3}{fake index3}
\item{v4}{fake index4}
}
}
\usage{
get_dup_test_data
}
\description{
A dataset containing several property keys such as policy number, policy effective date, premises number etc..
}
\keyword{datasets}
