% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_duplicates.R
\name{get_duplicates}
\alias{get_duplicates}
\title{Title}
\usage{
get_duplicates(indata, ...)
}
\arguments{
\item{indata}{dataframe}

\item{...}{keys that you want to test uniqueness}
}
\value{
a
}
\description{
Title
}
\examples{
get_duplicates(iris,Species)
}
